+++ shared/CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.0.0.45] - 2023-03-30
+
+- Add a deadline to the primary event handling to get visibility into concurrency issues
+
 ## [0.0.0.44] - 2023-03-28
 
 - Add UnmarshalJSON to KeyPair to unmarshal public key correctly