+++ shared/modules/doc/CHANGELOG.md
@@ -7,6 +7,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.0.0.10] - 2023-03-30
+
+- `Consensus` - improved documentation for supporting interfaces
+- `Consensus` - Consolidated `ResetRound`, `ResetForNewHeight` `ClearLeaderMessagesPool`
+- `Persistence` - Consolidated `Close` and `Release` and avoid returning an `error
+
 ## [0.0.0.9] - 2023-03-29
 
 - Improved README.md